more tidying, reducing duplicate code
