+++ src/renderer/rendererutils.ts
@@ -49,6 +49,7 @@ import {
 import { ConfigurationSchema } from 'config/configSchema';
 import { getLocalePhrase, Language } from 'localisation/translations';
 import { Phrase } from 'localisation/phrases';
+import { Dispatch, SetStateAction } from 'react';
 
 const getVideoResult = (video: RendererVideo): boolean => {
   return video.result;
@@ -1204,6 +1205,27 @@ const getVideoGroupIds = (
   );
 };
 
+const lockVideos = (
+  targets: Array<RendererVideo>,
+  lock: boolean,
+  setVideoState: Dispatch<SetStateAction<Array<RendererVideo>>>,
+) => {
+  const disk = targets.filter((v) => !v.cloud);
+  const cloud = targets.filter((v) => v.cloud);
+
+  const ipc = window.electron.ipcRenderer;
+  ipc.sendMessage('videoButtonDisk', ['protect', lock, disk]);
+  ipc.sendMessage('videoButtonCloud', ['protect', lock, cloud]);
+
+  setVideoState((prev) => {
+    return prev.map((rv) => {
+      return targets.some((target) => rv.uniqueId === target.uniqueId)
+        ? { ...rv, isProtected: lock }
+        : rv;
+    });
+  });
+};
+
 export {
   getFormattedDuration,
   getVideoResult,
@@ -1274,4 +1296,5 @@ export {
   getVideoParent,
   getVideoGroup,
   getVideoGroupIds,
+  lockVideos,
 };
